# openapi.yaml
openapi: "3.1.0"
info:
  title: IICP Directory Service
  version: "1.6.0"
  description: >
    IICP directory service — node registration, heartbeat, discovery, and
    node detail. Traffic NEVER routes through the directory; it is
    bootstrap and discovery only. See spec/iicp-dir.md for the normative
    sub-protocol specification.
  contact:
    url: https://iicp.network
  license:
    name: MIT
servers:
  - url: https://iicp.network/v1
    description: Production
...
paths:
...
  /dispatch/ticket:
    post:
      operationId: issueDispatchRouteTicket
      tags: [discovery]
      summary: Request one short-lived route-bearing dispatch ticket
      description: >
        Control-plane-only route selection. Task prompts and payloads are
        rejected. Clients may exclude previously failed node ID prefixes when
        requesting the next candidate. Full tickets and route URLs must not be
        logged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [intent]
              additionalProperties: false
              properties:
                intent: { type: string }
                region: { type: string }
                min_reputation: { type: number, minimum: 0, maximum: 1 }
                node_id: { type: string }
                node_id_prefix: { type: string, minLength: 4, maxLength: 36 }
                exclude_node_id_prefixes:
                  type: array
                  maxItems: 10
                  uniqueItems: true
                  items: { type: string, minLength: 4, maxLength: 36 }
      responses:
        "201":
          description: Ticketed concrete route; Cache-Control is no-store
        "404":
          description: No eligible route
        "409":
          description: Ambiguous node prefix
        "422":
          $ref: "#/components/responses/BadRequest"
        "503":
          description: Directory ticket signing is not configured
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: node_token issued at registration (Phase 1). JWT in Phase 2+.
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...